Restructured flashcard components to have one container
